backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.1.8
colorama==0.4.6
cryptography==46.0.5
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.124.4
fastapi-cli==0.0.18
fastapi-cloud-cli==0.7.0
fastar==0.9.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==3.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.2
PyMySQL==1.1.2
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.55.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.44.0
tomli==2.4.0
typer==0.20.1
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
watchfiles==0.24.0
websockets==13.1
